First cut vibrator control from Honey Select
This required 'porting' the Buttplug Core and Client
libraries into a form that would work on .net3.5 or more
accurately, the version of mono that Unity was using
when this game was built.

Most of the infrastructure is the same, although the
JSON support needed to be replaced. The current
implementation here is not perfect: most incoming
messages are now supported by only
SingleMotorVibrateCmd has been tested outbound.

The actual game integration watches for the sweet spot
indicator, then turns up the vibes to max. When it goes
the vibes switch off.